+++ src/index.js
@@ -1,7 +1,0 @@
-const { GatewayServer } = require('./gateway-server');
-const { OperationServer } = require('./operation-server');
-
-module.exports = {
-  GatewayServer,
-  OperationServer,
-};